+++ tests/data/test_import_datasets.py
@@ -128,5 +128,42 @@ class ImportDatasetsTest(MongoMockMixin, unittest.TestCase):
             self.assertEqual(dataset.chip_name, "IlluminaOvineSNP50")
 
 
+class DatasetsNotExistTest(MongoMockMixin, unittest.TestCase):
+    def setUp(self):
+        self.runner = CliRunner()
+        self.input_file = DATA_DIR / "test-dataset_not_exists.csv"
+
+    def tearDown(self):
+        Dataset.objects.delete()
+
+        super().tearDown()
+
+    @patch('src.features.smarterdb.Dataset.working_dir',
+           new_callable=PropertyMock)
+    @patch('src.data.import_datasets.get_raw_dir')
+    def test_import_datasets(self, my_raw_dir, my_working_dir):
+        # create a temporary directory using the context manager
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            working_dir = pathlib.Path(tmpdirname)
+            raw_dir = DATA_DIR / "raw"
+
+            # assign return value to mocked property
+            my_raw_dir.return_value = raw_dir
+            my_working_dir.return_value = working_dir
+
+            result = self.runner.invoke(
+                import_datasets,
+                [
+                    "--types",
+                    "genotype",
+                    "background",
+                    str(self.input_file),
+                ]
+            )
+
+            self.assertNotEqual(0, result.exit_code, msg=result.exception)
+            self.assertEqual(Dataset.objects.count(), 0)
+
+
 if __name__ == '__main__':
     unittest.main()